Refactor tests a little
